Refactor and cleanup. - Move core functions to ./lib/core.ts (calc, formatOutput, handleFlags, parse) - Clean index.ts - Rename parse to main - Add comments and tidy
