+++ lib/operators.ts
@@ -1,0 +1,6 @@
+export const operators: { [key: string]: Function } = {
+  "+": (a: number, b: number) => b + a,
+  "-": (a: number, b: number) => b - a,
+  "*": (a: number, b: number) => b * a,
+  "/": (a: number, b: number) => b / a
+};
